requirements.txt
ultralytics
roboflow
opencv-python
matplotlib
seaborn
pandas
mlflow
fastapi
uvicorn
python-multipart
sqlalchemy
jinja2
ipykernel
pyyaml
pydantic
requests